pyproject.toml
[project]
name = "wxabc"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "cryptography>=43.0.3",
  "loguru>=0.7.2",
  "pydantic>=2",
  "requests>=2.32.3",
  "aioredis>=2.0.1",
  "fastapi>=0.115.4",
  "loguru>=0.7.2",
  "opentelemetry-api>=1.28.1",
  "opentelemetry-distro>=0.49b1",
  "opentelemetry-exporter-otlp>=1.28.1",
  "opentelemetry-instrumentation-fastapi>=0.49b1",
  "opentelemetry-instrumentation-redis>=0.49b1",
  "opentelemetry-instrumentation-requests>=0.49b1",
  "opentelemetry-sdk>=1.28.1",
  "pydantic>=2",
  "redis>=5.2.0",
  "requests>=2.32.3",
  "uvicorn>=0.32.0",
]

[tool.uv]
index-url = "https://mirrors.tencent.com/pypi/simple/"
dev-dependencies = [
  "ipdb>=0.13.13",
  "pytest-asyncio>=0.24.0",
  "pytest>=8.3.2",
  "ruff>=0.6.4",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-s -vvv"
testpaths = ["tests"]

[project.optional-dependencies]
llm = ["langchain<3", "langchain-openai<0.3"]
dask = ["dask<=2024", "jpype1>=1.5.0"]
data = ["pyspark==3.5.3"]
aio = ["aiohttp<4", "aioresponses"]
